Replace old test cases
