+++ tools/run_pysdk_local_infinity_test.py
@@ -23,7 +23,7 @@ def python_sdk_test(python_test_dir: str, pytest_mark: str):
             "-x",
             "-m",
             pytest_mark,
-            f"{python_test_dir}/test/cases",
+            f"{python_test_dir}/test/internal_cases",
             "--local-infinity",
         ],
         stdout=sys.stdout,
